Fixed: (TEST) Avoid texture garbage generation code Fixed: (TEST) Avoid runtime texture space compression Fixed: (TEST) Reduced available texture mem size Fixed: Init MAXSORTTEX and MAXTPAGES with defined values (instead of hardcoded)

+++ plugins/peopsxgl/pgl_prim.c
@@ -1130,7 +1130,7 @@ BOOL CheckAgainstFrontScreen(short imageX0,short imageY0,short imageX1,short ima
 
 ////////////////////////////////////////////////////////////////////////
 
-void PrepareFullScreenUpload (int Position)
+void PrepareFullScreenUpload (short Position)
 {
  if (Position==-1)                                     // rgb24
   {
